app/auth/jwt_utils.py
"""
JWT creation and decoding utilities.
"""

from datetime import datetime, timedelta, timezone
from jose import jwt, JWTError
from app.core.config import settings


def create_jwt(payload: dict) -> str:
    expire = datetime.now(timezone.utc) + timedelta(
        minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
    )
    payload.update({"exp": expire})
    return jwt.encode(payload, settings.SECRET_KEY, algorithm="HS256")


def decode_jwt(token: str) -> dict | None:
    try:
        return jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
    except JWTError:
        return None
