requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.7
rq==1.16.2
python-jose==3.3.0
pydantic==2.9.0
pydantic-settings==2.3.4
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.21.0
requests==2.32.3
python-multipart==0.0.9
email-validator==2.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1